requirements.txt
Flask
opencv-python
pyserial
google-generativeai
requests